+++ src/features/home/ScheduleWidget/hooks/useMiniCalendarItemPosition.ts
@@ -1,0 +1,24 @@
+import { useState } from 'react';
+
+const useMiniCalendarItemPosition = () => {
+  const [indexPositions, setIndexPositions] = useState(
+    {} as { [key: number]: number },
+  );
+
+  const checkItemLayout = (
+    index: number,
+    event: {
+      nativeEvent: { layout: { x: number } };
+    },
+  ) => {
+    const x = event.nativeEvent.layout.x;
+    setIndexPositions(prev => ({ ...prev, [index]: x }));
+  };
+
+  return {
+    indexPositions,
+    checkItemLayout,
+  };
+};
+
+export default useMiniCalendarItemPosition;
